requirements.txt
torch~=2.1.0+cu118
numpy~=1.26.4
opencv-python~=4.8.1.78
pandas~=2.1.2
torchvision~=0.16.0+cu118
Pillow~=9.3.0
tqdm~=4.66.1
timm~=1.0.9
matplotlib~=3.8.1